% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atb_lexicon.R
\name{abbreviation_to_full_atb_name}
\alias{abbreviation_to_full_atb_name}
\title{Replace antibiotic abbreviations by full name}
\usage{
abbreviation_to_full_atb_name(string_to_change, language = "English")
}
\arguments{
\item{string_to_change}{Vector containing the antibiotic abbreviations}

\item{language}{Target manguage of the full antibiotic names, "English" and "French" 
are available. Defaults to "English"}
}
\description{
This function takes a vector of characters, and replace the antibiotic abbreviations
by their full name
}
\examples{
abbreviation_to_full_atb_name()
}
\keyword{abbreviations}
\keyword{antibiotic}
\keyword{names,}
