% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CB_OD_growth.R
\name{calculate_CB_OD_growth}
\alias{calculate_CB_OD_growth}
\title{Convert OD into growth percentage}
\usage{
calculate_CB_OD_growth(processed_OD_data_list, negative_control_col = 12,
  positive_control_col = 11, highest_conc_col.A = 1,
  lowest_conc_col.A = 10, highest_conc_row.B = 8, lowest_conc_row.B = 1,
  number_of_lines = 8)
}
\arguments{
\item{processed_OD_data_list}{List of data.frame. The first column of each
data.frame being the denomination of what is in the line and the following
columns being OD values read by the plate reader. The output of extract_MIC_OD_data
is already in the correct format.}

\item{negative_control_col}{Number of the column that contains the negative control.
Defaults to 12}

\item{positive_control_col}{Number of the column that contains the positive control.
Defaults to 11}

\item{lowest_conc_col.A}{Number of the column that contains the lowest concentration
of ATB.A Defaults to 10}

\item{lowest_conc_row.B}{Number of the row that contains the lowest concentration
of ATB.B Defaults to 1}

\item{number_of_lines}{Number of lines used by plate. Defaults to 8}

\item{higest_conc_col.A}{Number of the column that contains the highest concentration
of ATB.A. Defaults to 1}

\item{higest_conc_row.B}{Number of the row that contains the highest concentration
of ATB.B. Defaults to 8}
}
\description{
Calculate growth percentage in each well in order to determine the MIC of A and B for each row
}
\examples{
calculate_CB_OD_growth()
}
\keyword{CB,}
\keyword{OD,}
\keyword{growth}
\keyword{microdilution,}
