% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_clsi_breakpoints.R
\name{CB_clsi_breakpoints}
\alias{CB_clsi_breakpoints}
\title{Determine CLSI sensitivity}
\usage{
CB_clsi_breakpoints(CB.table, species = "Acinetobacter", MIC.A.column = 3,
  MIC.B.column = 4, FICI.conc.A.column = 6, FICI.conc.B.column = 7,
  ATB.A.column = 1, ATB.B.column = 2)
}
\arguments{
\item{CB.table}{data.frame having MIC.A and MIC.B in one column respectively.}

\item{species}{Species of the studied bacteria, for the moment only "Acinetobacter"
is available. Defaults to "Acinetobacter"}

\item{MIC.A.column}{ID of the column containing the MIC values in CB.table. Defaults to 2}

\item{MIC.B.column}{ID of the column containing the MIC values in CB.table. Defaults to 2}

\item{ATB.A.column}{ID of the column containing the ATB names in CB.table. Defaults to 1}

\item{ATB.B.column}{ID of the column containing the ATB names in CB.table. Defaults to 1}
}
\description{
This function given a data.frame with MICs will add a column to this table
giving the corresponding CLSI sensitivity (Sensitive, Intermediate or Resistant)
}
\examples{
CB_clsi_breakpoints()
}
\keyword{placeholder}
