% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucast_breakpoints.R
\name{eucast_breakpoints}
\alias{eucast_breakpoints}
\title{Determine EUCAST sensitivity}
\usage{
eucast_breakpoints(MIC.table, separator = " - ", species = "Acinetobacter",
  MIC.column = 2, ATB.column = 1)
}
\arguments{
\item{MIC.table}{data.frame having MIC in one column. If multiple MIC are given
in one row they must be separated by separator}

\item{separator}{When multiple MIC are given in one cell they must be separated by
separator. Defaults to " - "}

\item{species}{Species of the studied bacteria, for the moment only "Acinetobacter"
is available. Defaults to "Acinetobacter"}

\item{MIC.column}{ID of the column containing the MIC values in MIC.table. Defaults to 2}

\item{ATB.column}{ID of the column containing the ATB names in MIC.table. Defaults to 1}
}
\description{
This function given a data.frame with MIC will add a column to this table 
giving the corresponding EUCAST sensitivity (Sensitive, Intermediate or Resistant)
}
\examples{
eucast_breakpoints()
}
\keyword{placeholder}
