% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CB_OD_data.R
\name{extract_CB_OD_data}
\alias{extract_CB_OD_data}
\title{Extract data from spectrophotometer output}
\usage{
extract_CB_OD_data(raw_data_filename, max.conc.A, max.conc.B, startRow = 24,
  endRow = 32)
}
\arguments{
\item{raw_data_filename}{path to the excel file}

\item{max.conc.A}{vector giving the maximum concentration of
antibiotic A.The first element the vector
corresponds to the max concentration of antibiotic A for the first sheet.}

\item{max.conc.B}{vector giving the maximum concentration of
antibiotic B.The first element the vector
corresponds to the max concentration of antibiotic B for the first sheet.}

\item{startRow}{Excel row number where the OD readings start}

\item{endRow}{Excel row number where the OD readings stop}
}
\description{
This function allows you to extract optical density (OD) data from raw
output given by the TECAN OD reader, and store it to a list of
data.frame() (one data.frame per excel sheet)
}
\examples{
coucou
}
\keyword{CB,}
\keyword{Optical}
\keyword{density}
