% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MIC_OD_data.R
\name{extract_MIC_OD_data}
\alias{extract_MIC_OD_data}
\title{Extract data from spectrophotometer output}
\usage{
extract_MIC_OD_data(raw_data_filename, max.conc, startRow = 24, endRow = 32)
}
\arguments{
\item{raw_data_filename}{path to the excel file}

\item{max.conc}{list of vector giving the maximum concentration of
antibiotic used for each row of each plate. The first element of the list
corresponds to the first excel sheet. The first element of each vector
corresponds to the first line of the plate}

\item{startRow}{Excel row number where the OD readings start}

\item{endRow}{Excel row number where the OD readings stop}
}
\description{
This function allows you to extract optical density (OD) data from raw
output given by the TECAN OD reader, and store it to a list of
data.frame() (one data.frame per excel sheet)
}
\examples{
extract_MIC_OD_data("../../CahierDeLabo/MyELN/JPIAMR-Acineto/MIC/
MIC_26-09-17/MIC-26-09-17.xlsx",
list(c(rep(512,4),rep(128,4)),512,512,512,512,512))
}
\keyword{MIC,}
\keyword{Optical}
\keyword{density}
