% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_CB_non_OD_report.R
\name{generate_CB_non_OD_report}
\alias{generate_CB_non_OD_report}
\title{Generate a PDF with CB raw results}
\usage{
generate_CB_non_OD_report(title, date, data_file, species,
  negative_control_col = 12, positive_control_col = 11,
  highest_conc_col.A = 1, lowest_conc_col.A = 10, highest_conc_row.B = 8,
  lowest_conc_row.B = 1, number_of_lines = 8, knit_root_dir = getwd(),
  output_dir = getwd())
}
\arguments{
\item{title}{Title of the PDF file}

\item{data_file}{Path to the output from the plate reader}

\item{negative_control_col}{Number of the column that contains the negative control.
Defaults to 12}

\item{positive_control_col}{Number of the column that contains the positive control.
Defaults to 11}

\item{lowest_conc_col.A}{Number of the column that contains the lowest concentration
of ATB.A Defaults to 10}

\item{lowest_conc_row.B}{Number of the row that contains the lowest concentration
of ATB.B Defaults to 1}

\item{number_of_lines}{Number of lines used by plate. Defaults to 8}

\item{knit_root_dir}{Directory where the report should be knitted (using knitr)
Defaults to the current working directory}

\item{output_dir}{Directory where the PDF should be saved. Defaults to the current working directory}

\item{max.conc.A}{vector giving the maximum concentration of
antibiotic A.The first element the vector
corresponds to the max concentration of antibiotic A for the first sheet.}

\item{max.conc.B}{vector giving the maximum concentration of
antibiotic B.The first element the vector
corresponds to the max concentration of antibiotic B for the first sheet.}
}
\description{
Generate a PDF with CB raw results, with each plate and CB results
}
\examples{
generate_CB_non_OD_report()
}
\keyword{CB,}
\keyword{Report,}
\keyword{microdilution}
