% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MIC_report.R
\name{generate_MIC_report}
\alias{generate_MIC_report}
\title{Generate a PDF with MIC raw results}
\usage{
generate_MIC_report(title, date, max.conc, data_file, species,
  negative_control_col = 12, positive_control_col = 11,
  highest_conc_col = 1, lowest_conc_col = 10, number_of_lines = 8,
  startRow = 24, endRow = 32, knit_root_dir = getwd(),
  output_dir = getwd())
}
\arguments{
\item{title}{Title of the PDF file}

\item{max.conc}{list of vector giving the maximum concentration of
antibiotic used for each row of each plate. The first element of the list
corresponds to the first excel sheet. The first element of each vector
corresponds to the first line of the plates}

\item{data_file}{Path to the output from the plate reader}

\item{negative_control_col}{Number of the column that contains the negative control.
Defaults to 12}

\item{positive_control_col}{Number of the column that contains the positive control.
Defaults to 11}

\item{lowest_conc_col}{Number of the column that contains the lowest concentration
of tested antibiotic. Defaults to 10}

\item{number_of_lines}{Number of lines used by plate. Defaults to 8}

\item{knit_root_dir}{Directory where the report should be knitted (using knitr)
Defaults to the current working directory}

\item{output_dir}{Directory where the PDF should be saved. Defaults to the current working directory}
}
\description{
Generate a PDF with MIC raw results, with each plate and MIC results
}
\examples{
generate_MIC_report()
}
\keyword{MIC,}
\keyword{Report,}
\keyword{microdilution}
