% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomSeries.R
\name{geomSeries}
\alias{geomSeries}
\title{Geometric series}
\usage{
geomSeries(base, max)
}
\arguments{
\item{base}{Base of the series}

\item{max}{The maximum value of the series}
}
\description{
A function that given a base and a maximum value will calculate all the terms of 
a geometrical series
}
\examples{
geomSeries(2,512)
}
\keyword{geometric}
\keyword{sries}
