% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CB_OD.R
\name{plot_CB_OD}
\alias{plot_CB_OD}
\title{Generate plots from OD results}
\usage{
plot_CB_OD(processed_growth_data_list)
}
\arguments{
\item{processed_growth_data_list}{A list of data.frames as produced by the calculate_CB_OD_growth() function}
}
\description{
This function generates a plot for each plate read by OD. Each well is represented 
and the higher the growthition is the darker the well will be represented
}
\examples{
plot_CB_OD()

}
\keyword{CB,}
\keyword{OD,}
\keyword{growthition}
\keyword{plots,}
