% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MIC_OD.R
\name{plot_MIC_OD}
\alias{plot_MIC_OD}
\title{Generate plots from OD results}
\usage{
plot_MIC_OD(processed_OD_data_list)
}
\arguments{
\item{processed_OD_data_list}{A list of data.frames as produced by the extract_MIC_OD_data() function}
}
\description{
This function generates a plot for each plate read by OD. Each well is represented 
and the higher the OD is the darker the well will be represented
}
\examples{
plot_MIC_OD()
}
\keyword{MIC,}
\keyword{OD}
\keyword{plots,}
