% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CB_growth.R
\name{summary_CB_growth}
\alias{summary_CB_growth}
\title{Summarize CB results}
\usage{
summary_CB_growth(growth_data_list, number_of_lines = 8,
  species = "Acinetobacter")
}
\arguments{
\item{growth_data_list}{A list of data.frames returned by the calculate_CB_OD_growth() function}

\item{number_of_lines}{Number of lines used by plate. Defaults to 8}

\item{species}{Species of the studied bacteria, for the moment only "Acinetobacter"
is available. Defaults to "Acinetobacter"}
}
\description{
Given OD MIC data processed through calculate_CB_OD_growth() function creates
a data.frame with the CB results for each plate and a list of data.frames splitting the first data.frames by strain
(second element returned)
}
\examples{
summary_CB_growth()
}
\keyword{MIC,}
\keyword{results}
\keyword{summary,}
