% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_MIC_inhib.R
\name{summary_MIC_inhib}
\alias{summary_MIC_inhib}
\title{Summarize MIC results}
\usage{
summary_MIC_inhib(inhib_data_list, number_of_lines = 8,
  species = "Acinetobacter")
}
\arguments{
\item{inhib_data_list}{A list of data.frames returned by the calculate_MIC_OD_inhib() function}

\item{number_of_lines}{Number of lines used by plate. Defaults to 8}

\item{species}{Species of the studied bacteria, for the moment only "Acinetobacter"
is available. Defaults to "Acinetobacter"}
}
\description{
Given OD MIC data processed through calculate_MIC_OD_inhib() function creates
a data.frame with the MIC for each row and the row identifier (first element 
returned) and a list of data.frames splitting the first data.frames by strain
(second element returned)
}
\examples{
summary_MIC_inhib()
}
\keyword{MIC,}
\keyword{results}
\keyword{summary,}
